+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # ilert-rust CHANGELOG
 
+## 2023-05-12, Version 3.0.0
+
+* updated dependencies
+* fixed a few occurrences of `iLert` -> `ilert` due to rebranding, however decided to keep core structs on `ILert..` as `Ilert` is not much better
+* added endpoints to handle `/api/alerts`
+* added endpoints to handle `/api/incidents`
+* added endpoints to handle `/api/services`
+* added traits to support **PUT** and **DELETE** operations in builder style
+* added skip() and limit() options to .get() builders
+* created client.create() in favor of client.post(); `post()` is now deprecated
+* set default timeout to 25 seconds, was 10 seconds
+* **BREAKING** renamed `ScheduleApiResource` to `ScheduleGetApiResource`
+* **BREAKING** renamed `UserApiResource` to `UserGetApiResource`
+
 ## 2021-11-01, Version 2.0.0
 
 * **BREAKING** adjusted client to the new iLert API version `/api/v1` -> `/api`